忘记Int的Bug

+++ Emotion Diary/EmotionDiary.swift
@@ -93,10 +93,10 @@ class EmotionDiary: NSObject, NSCoding {
     
     class func getDiaryOfDay(date: NSDate) -> [EmotionDiary] {
         
-        let thisDay = date.timeIntervalSince1970 / (24 * 3600)
+        let thisDay = Int(date.timeIntervalSince1970 / (24 * 3600)
         let emotionDiaries = EmotionDiary.getEmotionDiariesFromStore()
         let thatDayDiary = emotionDiaries.filter { (diary) -> Bool in
-            let thatDay = diary.date.timeIntervalSince1970 / (24 * 3600)
+            let thatDay = Int(diary.date.timeIntervalSince1970 / (24 * 3600))
             return (thatDay == thisDay)
         }
         return thatDayDiary
